Create separate module for utils that need Android functionality
