+++ common/core-android/build.gradle
@@ -1,0 +1,43 @@
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+}
+
+android {
+    namespace 'com.alvindizon.panahon.core.android'
+    compileSdk buildVersions.compileSdkVersion
+
+    defaultConfig {
+        minSdk buildVersions.minSdkVersion
+        targetSdk buildVersions.targetSdkVersion
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+}
+
+
+dependencies {
+    implementation deps.androidX.coreKtx
+
+    // for Observer
+    implementation deps.androidX.lifecycleRuntimeKtx
+    // for savedstatehandle
+    implementation deps.androidX.activityCompose
+
+    // Coroutines
+    implementation deps.coroutines.core
+    implementation deps.coroutines.android
+
+    api deps.timber
+}
