+++ common/location/build.gradle
@@ -55,6 +55,7 @@ android {
 dependencies {
     implementation project(":common:design")
     implementation project(":common:core")
+    implementation project(":common:core-android")
 
     // Coroutines
     implementation deps.coroutines.core
