+++ common/navigation/build.gradle
@@ -50,6 +50,7 @@ android {
 
 dependencies {
     implementation project(":common:core")
+    implementation project(":common:core-android")
 
     // Coroutines
     implementation deps.coroutines.core
