+++ feature/home/build.gradle
@@ -58,6 +58,7 @@ android {
 dependencies {
     implementation project(":common:design")
     implementation project(":common:core")
+    implementation project(":common:core-android")
     implementation project(":common:location")
     implementation project(":common:navigation")
     implementation platform(deps.androidX.compose.bom)
