+++ settings.gradle
@@ -33,3 +33,4 @@ include ':feature:home'
 include ':common:navigation'
 include ':common:preferences'
 include ':feature:settings'
+include ':common:core-android'
